build(deps): Upgrade various dependencies
We have a critical security update on Docker auth, which is an
indirect dependency of the test code (seemingly via oras).

This is the culmination of different attempts at upgrading without
having to make substantial changes to the sloth source code.

Some notes:
* Upgrading go-helm-template (also made by slok) in test ultimately
  requires at least Go 1.22.
* The GH actions uses the root go.mod Go version, so these need
  to be synced.
* The next version of reloader, also made by slok, requires
  1.23, so both root and test needs to be on 1.23.
* Updated the dev and prod Dockerfiles to match the go version.
* Upgraded most dependencies to get as far as possible.
* Keeping Prometheus-related deps as certain newer versions
  don't work with the code. There might be further version jumps
  we can do here.

This results in errors on lint which are addressed in #35.