build(deps): Upgrade go-helm-template and indirect deps
This package is also made by slok and seemingly more actively
maintained. Both two direct deps are upgraded, presumably due to
to increased go version requirement.

The hope is that this updates deps in a more coherent and functional
manner compared to #13. This is necessary to address a Critical
security issue.